Allow duplicate instances for unique indexes
An index on an entity defines how an entity is unique. For a file this is for
example the host it is deployed on and the file path. This commit allows
configuration models where an entity is defined twice, IF all of its
attributes have the same value. The main use case for this is definening the
state of a package or service in multiple entities.

When a duplicate definition is detected in the index, it will add all new
attributes of the second instance to the first instance. It wil also
replace all references to the second instance with references to the
first instance. If a value is set on both instances, they need to be
identical.

Plugins that maintain state, for example a list of defined packages per host,
need to handle de-duplication. On "de-duplication", the attribute
__replaced_by__ is set on the second instance. The value of this attribute
is a reference to the first instance.