Fix the use of optional values
A relation with a multiplicity of [0:1] does not define a list but defines an attribute of which the
value is optional. When this is used in an implements statement, std::isset could be used but this
was problematic and only worked in specific cases.

This patch introduces the "is defined" test, as available in jinja. So the following implement
statement is valid:

implement Test using testImpl when self.attribute is defined