initial fix for improved provides tracking
introduced a promise object, that represents the promise to set a value
on a result variable. As long as a promise is outstanding, a list can
not be frozen

we are now mixing provides tracking based on counts and based on
promises. Some cleanup might be good, look in

ExecutionUnit and Instance.set_attribute