Fix a boundary condition in protocol client (#837)
When a call times out with the protocol client, it will throw an
exception instead of a returning the error code. This is caused by the
http error that is generated by the library and not the server.

A request_timeout option is introduced on transport setting to be able
to configure the timeout instead of the default 120s. This setting is
used by the unit test for this fix.