Issue/2500 agent api paging (#2521)
Implemented `get_list_paged` as an alternative to `get_list`.  While `get_list` uses `offset` and `limit` to select a section of output, `get_list_paged` relies on a combination of `start`, `end` (applied on the column we are basing the sorting on) and `limit`.  This method is now used by the following routes: 
 - `GET /agent`
 - `GET /agentproc`
 - `GET /compilereport` (this already did pagination, but the method used is now common to the others)

Co-authored-by: Guillaume Everarts de Velp <guillaume.everartsdevelp@inmanta.com>
Co-authored-by: arnaudsjs <2684622+arnaudsjs@users.noreply.github.com>