Extend programmatic API documentation for parts used by inmanta-sphinx (PR #2854)
# Description

Extend modules API and documentation to provide inmanta-sphinx with a stable API (see inmanta/inmanta-sphinx#43). This pull request doesn't add full documentation for all parts used by inmanta-sphinx. It mainly contains documentation for the parts that caused the docs builds to fail with a few others thrown in. I've created #2853 for the rest since it turned out to be a lot more than initially expected.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design