Bump pydantic dependency to 1.7.3 (PR #2957)
# Description

pytest-docker fails for the combination (os=fedora-latest, branch=iso3), because it uses Python3.9. The problem is situated in the pydantic dependency. Pydantic only support Python3.9 from version `1.7.0`, while inmanta-core still uses version `1.6.1`.

Next to that, there is an inconsistency between inmanta-core and the inmanta-lsm extension with respect to the pydantic dependency. Inmanta-core pins pydantic to version `1.6.1`, while inmanta-lsm pins it to version `1.7.3`. For that reason I decided bump the pydantic version to `1.7.3` on inmanta-core as well. This fixes the inconsistency and the failure on pytest-docker.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [ ] ~~Type annotations are present~~
- [ ] ~~Code is clear and sufficiently documented~~
- [ ] ~~No (preventable) type errors~~
- [ ] ~~Sufficient test cases (reproduces the bug/tests the requested feature)~~
- [ ] ~~Correct, in line with design~~
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design