Fix broken test case `test_resource_evolution` (PR #2966)
# Description

The `test_resource_evolution` test case broke due to the update to the std module that sets the `purge_on_delete` flag to false by default. The branches `merge-tool/2959/iso4` and `issue/2958-set-purge-on-delete-to-false-iso3` which should merge #2959 into the ISO4 and ISO3 respectively are also stuck on this issue. I will apply this change on those branches as well.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [ ] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design