Added modules v2 documentation for dev-to-prod workflow. (Issue #3039, PR #3058)
# Description

This is still a work in progress, though I think the development part is mostly done (some minor changes might be needed related to the `poetry install` editable discussion). I though it might be good to open this part up for review already because as we've noticed this might spark new discussions. As a result it's probably a good idea to keep the feedback loop relatively small.

closes #3039

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design