Added parsing and internal representation of modules v2 metadata (Issue #3106, PR #3107)
# Description

Adds parsing and internal representation of modules v2 metadata. Some changes to the `Module` class were required to fit this in the existing implementation. I tried to keep this as minimal as possible, which means this pull request leaves core in the state where it has entry points to parse V2 modules, but they are never used outside testing. `ModuleTool` and `Project` always work on `ModuleV1`. Further changes to `Project`, `Module` and its subclasses will have to be implemented by follow-up tickets.

I've created #3114 as a follow-up to make sure the documentation and usage of `Module` is updated.

(opened in a bit of a rush, I hope I haven't missed anything)

closes #3106

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: #3114)

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design