Prevent that the tests checkout the std module in the test/data/plugins_project/libs directory. (PR #3147)
# Description

The `test_code_manager` test case, executed `project.load()` in the `test/data/plugins_project` directory of the inmanta-core project. This way the std module got checked out into the `test/data/plugins_project/libs` directory. When the std module is present in the `test/data/plugins_project/libs` directory, the following error occurs when the test suite is ran:

```
import file mismatch:
imported module 'test_agent' has this __file__ attribute:
  /tmp/inmanta-core/tests/test_agent.py
which is not the same as the test file we want to collect:
  /tmp/inmanta-core/tests/data/plugins_project/libs/std/tests/test_agent.py
HINT: remove __pycache__ / .pyc files and/or use a unique basename for your test file modules

```

This PR makes sure that the `test_code_manager` test case executes `project.load()` in a temporary directory to prevent the above-mentioned issues.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design