Clean up the use of the pg fixtures (PR #3270)
# Description

This PR cleans up the postgres fixture to allow the use of an external database service when --postgresql-host is explicitly set. The biggest change is actually passing the credentials everywhere.

This was an experiment to see if we could speed up the tests using a single database and running them in parallel. This makes many test workers hang. This PR is the useful part of this experiment.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design