Fixed unstable tests by clearing build.env cache (Issue #3290, PR #3296)
# Description

Fixes some unstable tests by clearing `build.env._should_use_virtualenv` cache before and after yielding isolated fixture. A potential issue with this approach is that it accesses a private method, but since we only do so in our tests I think that should be acceptable.

closes #3290

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] ~~Type annotations are present~~
- [x] Code is clear and sufficiently documented
- [x] ~~No (preventable) type errors (check using make mypy or make mypy-diff)~~
- [x] ~~Sufficient test cases (reproduces the bug/tests the requested feature)~~
- [x] Correct, in line with design
- [x] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design