Added a generic fixture for db migration tests that can be used by core and extensions. (PR #3299)
# Description

Added a generic fixture for db migration tests. We'd discussed this on slack a couple of weeks ago. Now that we needed a new db migration test I thought it would be a good time to give this a try.
Let me know what you think (this is more a suggestion than anything else). The parametrization of fixtures isn't super clean, so you might not like it. Overall I'm of the opinion that it's an improvement over the fixture-per-test approach we had up till now. Another approach could be to just have a helper method instead of a fixture but I believe the fixture's advantage outweigh the disadvantages (simple setup and teardown, which would require more care when calling a helper method).

Have a look at inmanta/inmanta-lsm#687 to see this in use.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] ~~Sufficient test cases (reproduces the bug/tests the requested feature)~~
- [x] Correct, in line with design
- [x] End user documentation is included

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design