Restrict check on number of agent processes to agent processes only (PR #3323)
# Description

Some test verify whether an agent process is running or not. This check creates a diff between all the subprocesses of the process running the test between moment A and moment B. The problem is that PostgresSQL spawns subprocesses during the test case. As such, the diff is done incorrectly. This PR makes sure that the process diff considers Inmanta agent processes only.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design