Remove the init_namespace() method in favor of using the notify_change() method (PR #3328)
# Description

While trying the find out why the `ActiveEnv.init_namespace()` method was required, I noticed it would be better to make this logic part of the `ActiveEnv.notify_change()` method. In the previous implementation `init_namespace()` has to be called manually from different locations and in a specific order. The new implementation hides all the complexity in the `ActiveEnv` class.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design