Modified pth file to skip already loaded paths (PR #3354)
# Description

The pth file written by `inmanta.env.VirtualEnv` loads all paths from the parent venv. This pull request modifies it to only load those paths that are not already in `sys.path`. This fixes the test performance regression and prevents infinite recursion when the project venv is set to the encapsulating venv.

Do you think this requires an explicit test case or do we consider the fact that no infinite recursion occurs sufficient?

closes #3344
closes #3346

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design