Fixed DefineImport location inserted when autostd=True (PR #3394)
# Description

Attaches a valid location to `DefineImport` statement created by module when `autostd=True` similar to the one created by the project. While an error might not arise from this on iso4, I believe it's still best to patch it there as well and set a valid location.

I don't think it's relevant to add a test case for this. Let me know if you disagree.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design