removed newline from v2 test module description (PR #3438)
# Description

There was a newline in one of our test modules' description field. Setuptools does not support this and if I understand correctly has had a warning about this for some time. In the latest release they converted this warning to a `ValueError` (https://github.com/pypa/setuptools/blob/main/CHANGES.rst#v5900).

This warning seems to have been hidden from us (not confirmed). I'm adding this to #3388 to make sure similar warnings are reported correctly in the future.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] ~~Type annotations are present~~
- [x] ~~Code is clear and sufficiently documented~~
- [x] ~~No (preventable) type errors (check using make mypy or make mypy-diff)~~
- [x] ~~Sufficient test cases (reproduces the bug/tests the requested feature)~~
- [x] ~~Correct, in line with design~~
- [x] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design