Fixed project installation when the project has a requirement on a module that is not loaded by the model (PR #3557)
# Description

I discovered when modifying a project I had previously added requirements to that there was the following behavior. If a project has a module in the `requires` field of its config file but that module is not loaded by the model, the verification logic would fail with a `KeyError`. This is definitely not desired. I opted to resolve it by ignoring such modules for verification. Another approach would be to discover the installed version dynamically in such a case (instead of getting it out of `self.modules`) but that would be more complex and I'm not sure we really gain anything.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design