Raise a warning when trying to use repo of type package. (Issue #3565, PR #3638)
# Description

Modules v2 has introduced multiple repo types. Previously all repos were of type "git", while for iso5+ repos of type "package" are possible as well. The project template will include one such repo. This means iso4 needs to be compatible with both the new format and the new type (only in the sense that it should not break, iso4 should not actually use repos of type "package"). Iso4 should raise a warning for repos of type "package" but otherwise ignore them, and accept explicit specification of the git type, instead of failing on the new syntax.

issue #3565

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design