Ensure processes forked by Inmanta commands load the same config folder as their parent process (Issue #3765, PR #3766)
# Description

Ensure every command forked inherits the config dir of its parent

closes  #3765

This can be tested by putting an invalid file in `/etc/inmanta.d/` but that is outside the reach of the testsuite.
So there is no specific testcase for this.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )