Optimize resource list query (PR #3835)
# Description

The optimized query is based on https://stackoverflow.com/a/25536748, 1.a variant. I also added this link to the source

With the changes the average execution time on the same data (and database version) goes from the 500-600 ms range (https://explain.dalibo.com/plan/uWK#plan) to the 100-200 ms range (https://explain.dalibo.com/plan/yS8).
Note: These results are based on the query plan analyzer and not on extensive performance testing.

closes #3374

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] ~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design