Fixed an instance of nondeterministic behavior in the compiler (PR #3920)
# Description

Removes a source of nondeterminism from the compiler: construction of an `Instance` instance used a set to prune duplicates from a sequence. Iteration over this set was then indirectly used to drive control flow (a method was called that as a side effect appended to the execution queues). This PR performs the pruning in a deterministic way and adds a test case for a model that was affected by this.

closes #3034

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design