Remove shift/reduce conflicts by enforcing precedence rules (PR #3939)
# Description
This Commit removes all the warnings of shift/reduce conflicts by enforcing a shift.
By default if no precedence rules are found, a shift is performed and a warning issued.
By adding the precedence rules to perform a shift when there are conflict the same behavior
should be kept without the warnings.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design