Expressions are now treated as statements (Issue #3367, PR #3959)
# Description

This commit makes it so that all expressions are now treated as statements.
Some changes where made to MLS as well as it can now occur as a "statement string" or a multi-line comment,
where previously a standalone MLS was always a comment. To be a string it needed to be part of  an assignment.
If the MLS is at the top of a file it will be considered as a comment, otherwise as a Literal.

closes #3367

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design