Fix bug where `inmanta project install` and `inmanta project update` always invokes pip, even when all dependencies are already met. (Issue #4055, PR #4243)
# Description

The `are_installed()` method in `env.py`, that checks whether a requirement is installed in the active Python environment, didn't take into account the markers set on a requirement. As such, the method would always evaluate to `False` when a marker was set that wasn't applicable for the active environment. This caused degraded performance on all test suites that ran using python >=3.7 and used the `project` fixture, because the requirements.txt file of the std module has the following requirement set:

```
dataclasses~=0.7;python_version<'3.7'
```

Each time the `are_installed()` method evaluates to False, a pip sub-process is spawned. This slowed down the tests a lot. This patch increases the runtime of the `lsm` test suite by 2.2x

closes #4055

# Self Check:

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design