Don't re-raise TimeoutError on close connection pool (PR #4352)
# Description

The test suite sets the timeout to close the connection pool to the database to 3 seconds. This doesn't seem to be enough for the test cases that run the database migration script. As such, these test cases often fail with a TimeoutError.  I don't think there is reason why we have to re-raise the TimeoutError. All connection are forcefully closed when a timeout occurs, so I think that just writing a log message is enough here.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design