Add endpoint for partial compiles (Issue #4271, PR #4348)
# Description

This commit adds an extra API endpoint to the server that is invoked to store the resource of an partial compile. This new endpoint is equivalent to the put_version endpoint with the following two changes:

The resources argument only contains the resources that resulted from the partial compile.
The endpoint gets an extra argument that indicates which resource sets should be deleted from the model.

closes #4271

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design