Improved transparency of get_connection method in inmanta.data (PR #4434)
# Description

Suggestion for improvement of the `get_connection` method in `inmanta.data.BaseDocument`. This class is part of the stable API and I changed the method signature in the sense that it can now only be used with a context manager where previously it could also be just awaited. But I'm pretty sure that this method isn't used outside of core, and if it is I expect it's used as a context manager.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design